# server/specs/openapi.yaml
openapi: 3.1.0
info:
  title: LoseIt API
  version: "0.1.0"
  license:
    name: MIT OR Apache-2.0
    identifier: MIT OR Apache-2.0
  description: |
    HTTP API for the LoseIt server. All routes (except `/health`) require a
    bearer token in the `Authorization` header. Token validation is delegated
    to a pluggable authenticator; the first identity seen for a given
    `(issuer, external_id)` pair is auto-provisioned into a local user row.

    All paths are served under the `/api/v1` prefix.

    Decimal-valued fields (weights, nutrition, quantities) are transported as
    JSON numbers but produced/consumed by the server as arbitrary-precision
    decimals — clients should avoid binary float round-trips for fields like
    `weight_kg`, `grams`, and `quantity`.

    ## v1 pre-release wire-shape changes

    The following endpoints returned bare JSON arrays in early drafts and
    now return paginated envelopes (`{ results, total, limit, offset }`):

    * `GET /log`
    * `GET /weights`

    Both endpoints' `from` and `to` query parameters were previously
    required and are now optional. With no parameters, the response is
    the 100 most-recent entries across the full history.

    The `GET /foods/search` response schema migrated from the bespoke
    `FoodSearchPage` shape to the unified `PaginatedFoodSearchHits`
    envelope; field names are unchanged.
servers:
  - url: "https://{host}/api/v1"
...
    variables:
      host:
        default: api.loseit.invalid
...
security:
  - bearerAuth: []
...
paths:
...
  /foods/search:
    get:
      tags: [Foods]
      operationId: search_foods
      summary: Search foods
      description: |
        Full-text search across the foods catalogue (OFF, USDA, and the
        caller's custom foods). A blank `q` returns `400 bad_request`.

        Default `limit` is 100; the server silently clamps requests above
        500. Sentinel foods used by `POST /log/quick_add` are excluded
        from results.
      parameters:
        - in: query
          name: q
          required: true
          schema: { type: string, minLength: 1 }
          description: Search query string.
        - in: query
          name: limit
          schema: { type: integer, format: int64, minimum: 0 }
          description: Page size. Server applies a cap if exceeded.
        - in: query
          name: offset
          schema: { type: integer, format: int64, minimum: 0 }
          description: Page offset, starting at 0.
      responses:
        "200":
          description: Search page.
          content:
            application/json:
              schema: { $ref: "#/components/schemas/PaginatedFoodSearchHits" }
        "400": { $ref: "#/components/responses/BadRequest" }
        "401": { $ref: "#/components/responses/Unauthorized" }
  /foods/mine:
    get:
      tags: [Foods]
      operationId: list_my_foods
      summary: List the caller's custom foods
      description: |
        Lists `user`-source foods owned by the caller, optionally filtered
        by a case-insensitive substring match against `name` and `brands`.
        The sentinel `__quick_add__` food (used by `POST /log/quick_add`)
        is never returned.

        Default `limit` is 100; server silently clamps requests above 500.
      parameters:
        - in: query
          name: q
          schema: { type: string, maxLength: 200 }
          description: Substring filter; empty/whitespace is treated as absent.
        - in: query
          name: limit
          schema: { type: integer, format: int64, minimum: 0 }
          description: Page size; 0 or omitted → 100.
        - in: query
          name: offset
          schema: { type: integer, format: int64, minimum: 0 }
          description: Page offset, starting at 0.
      responses:
        "200":
          description: Page of the caller's custom foods.
          content:
            application/json:
              schema: { $ref: "#/components/schemas/PaginatedFoodSearchHits" }
        "400": { $ref: "#/components/responses/BadRequest" }
        "401": { $ref: "#/components/responses/Unauthorized" }
  /foods/recent:
    get:
      tags: [Foods]
      operationId: list_recent_foods
      summary: Foods the caller has logged most recently
      parameters:
        - in: query
          name: limit
          schema: { type: integer, format: int64, minimum: 1 }
          description: Maximum number of foods to return.
      responses:
        "200":
          description: Recent foods.
          content:
            application/json:
              schema:
                type: array
                items: { $ref: "#/components/schemas/FoodSearchHit" }
        "400": { $ref: "#/components/responses/BadRequest" }
        "401": { $ref: "#/components/responses/Unauthorized" }
  /foods/frequent:
    get:
      tags: [Foods]
      operationId: list_frequent_foods
      summary: Foods the caller logs most frequently
      parameters:
        - in: query
          name: limit
          schema: { type: integer, format: int64, minimum: 1 }
          description: Maximum number of foods to return.
      responses:
        "200":
          description: Frequent foods.
          content:
            application/json:
              schema:
                type: array
                items: { $ref: "#/components/schemas/FoodSearchHit" }
        "400": { $ref: "#/components/responses/BadRequest" }
        "401": { $ref: "#/components/responses/Unauthorized" }
  /foods/barcode/{barcode}:
    get:
      tags: [Foods]
      operationId: get_food_by_barcode
      summary: Look up a food by barcode
      parameters:
        - in: path
          name: barcode
          required: true
          schema: { type: string }
          description: Barcode string (EAN/UPC).
      responses:
        "200":
          description: Food detail.
          content:
            application/json:
              schema: { $ref: "#/components/schemas/FoodDetail" }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "404": { $ref: "#/components/responses/NotFound" }
  /foods:
    post:
      tags: [Foods]
      operationId: create_food
      summary: Create a custom food
      description: |
        Creates a `user`-source food owned by the caller. A default
        100 g system serving is auto-seeded.
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/FoodCreate" }
      responses:
        "201":
          description: Created.
          content:
            application/json:
              schema: { $ref: "#/components/schemas/FoodDetail" }
        "400": { $ref: "#/components/responses/BadRequest" }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "409": { $ref: "#/components/responses/Conflict" }
  /foods/{id}:
    get:
      tags: [Foods]
      operationId: get_food
      summary: Get a food by id, including its servings
      parameters:
        - $ref: "#/components/parameters/PathId"
      responses:
        "200":
          description: Food detail.
          content:
            application/json:
              schema: { $ref: "#/components/schemas/FoodDetail" }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "404": { $ref: "#/components/responses/NotFound" }
    patch:
      tags: [Foods]
      operationId: update_food
      summary: Update a custom food
      description: Only the owner of a `user`-source food may patch it.
      parameters:
        - $ref: "#/components/parameters/PathId"
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/FoodPatch" }
      responses:
        "200":
          description: Updated food.
          content:
            application/json:
              schema: { $ref: "#/components/schemas/FoodDetail" }
        "400": { $ref: "#/components/responses/BadRequest" }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "403": { $ref: "#/components/responses/Forbidden" }
        "404": { $ref: "#/components/responses/NotFound" }
    delete:
      tags: [Foods]
      operationId: delete_food
      summary: Delete a custom food
      description: |
        Returns `403` if the food isn't owned by the caller and `409` if
        the food is referenced by existing log entries.
      parameters:
        - $ref: "#/components/parameters/PathId"
      responses:
        "204": { description: Deleted. }
        "400": { $ref: "#/components/responses/BadRequest" }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "403": { $ref: "#/components/responses/Forbidden" }
        "404": { $ref: "#/components/responses/NotFound" }
        "409": { $ref: "#/components/responses/Conflict" }
  /foods/{food_id}/servings:
    post:
      tags: [Servings]
      operationId: create_serving
      summary: Create a serving on a food
      parameters:
        - in: path
          name: food_id
          required: true
          schema: { type: string, format: uuid }
          description: Id of the food the serving belongs to.
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/ServingCreate" }
      responses:
        "201":
          description: Created.
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Serving" }
        "400": { $ref: "#/components/responses/BadRequest" }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "403": { $ref: "#/components/responses/Forbidden" }
        "404": { $ref: "#/components/responses/NotFound" }
  /servings/{id}:
    patch:
      tags: [Servings]
      operationId: update_serving
      summary: Update a serving
      parameters:
        - $ref: "#/components/parameters/PathId"
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/ServingPatch" }
      responses:
        "200":
          description: Updated serving.
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Serving" }
        "400": { $ref: "#/components/responses/BadRequest" }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "403": { $ref: "#/components/responses/Forbidden" }
        "404": { $ref: "#/components/responses/NotFound" }
    delete:
      tags: [Servings]
      operationId: delete_serving
      summary: Delete a serving
      parameters:
        - $ref: "#/components/parameters/PathId"
      responses:
        "204": { description: Deleted. }
        "400": { $ref: "#/components/responses/BadRequest" }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "403": { $ref: "#/components/responses/Forbidden" }
        "404": { $ref: "#/components/responses/NotFound" }
        "409": { $ref: "#/components/responses/Conflict" }
  /servings/{id}/default:
    post:
      tags: [Servings]
      operationId: set_default_serving
      summary: Make this serving the default for its food
      description: |
        Atomically flips the `is_default` flag: the named serving becomes
        the food's default and any previous default is unset.
      parameters:
        - $ref: "#/components/parameters/PathId"
      responses:
        "200":
          description: The serving in its post-flip state.
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Serving" }
        "400": { $ref: "#/components/responses/BadRequest" }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "403": { $ref: "#/components/responses/Forbidden" }
        "404": { $ref: "#/components/responses/NotFound" }
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: |
        Bearer token validated by the configured authenticator. In
        development a static `DEV_AUTH_BYPASS` token may be used; in
        production tokens are expected to be JWTs validated against a JWKS
        endpoint.
  parameters:
    PathId:
      in: path
      name: id
      required: true
      schema: { type: string, format: uuid }
      description: Resource identifier.
  responses:
    BadRequest:
      description: Request was malformed or failed domain validation.
      content:
        application/json:
          schema: { $ref: "#/components/schemas/Error" }
    Unauthorized:
      description: Missing or invalid credentials.
      content:
        application/json:
          schema: { $ref: "#/components/schemas/Error" }
    Forbidden:
      description: Authenticated but not allowed to perform this action.
      content:
        application/json:
          schema: { $ref: "#/components/schemas/Error" }
    NotFound:
      description: Resource does not exist or is not visible to the caller.
      content:
        application/json:
          schema: { $ref: "#/components/schemas/Error" }
    Conflict:
      description: Domain invariant prevents the operation.
      content:
        application/json:
          schema: { $ref: "#/components/schemas/Error" }
  schemas:
    Error:
      type: object
      required: [code, message]
      properties:
        code:
          type: string
          description: Stable, machine-readable error identifier.
          examples:
            - bad_request
            - unauthorized
            - forbidden
            - not_found
            - conflict
            - upstream_error
            - internal_error
        message:
          type: string
    Decimal:
      description: Arbitrary-precision decimal serialized as a JSON number.
      type: number
    NullableDecimal:
      description: Arbitrary-precision decimal or `null`.
      type: [number, "null"]
...
    FoodSource:
      type: string
      enum: [off, user, usda]
...
    NutriscoreGrade:
      type: string
      enum: [a, b, c, d, e]
...
    NutritionPer100g:
      type: object
      description: Per-100g nutrition. All fields are optional/nullable.
      properties:
        energy_kcal: { $ref: "#/components/schemas/Decimal" }
        protein_g: { $ref: "#/components/schemas/Decimal" }
        carbs_g: { $ref: "#/components/schemas/Decimal" }
        fat_g: { $ref: "#/components/schemas/Decimal" }
        fiber_g: { $ref: "#/components/schemas/Decimal" }
        sugar_g: { $ref: "#/components/schemas/Decimal" }
        sodium_g:
          allOf:
            - $ref: "#/components/schemas/Decimal"
          description: |
            Sodium in grams per 100 g (OFF convention). Clients should
            multiply by 1000 to display as milligrams, which is the
            customer-expected unit.
        saturated_fat_g: { $ref: "#/components/schemas/Decimal" }
    Serving:
      type: object
      required: [id, label, grams, is_default, source, sort_order]
      properties:
        id: { type: string, format: uuid }
        label: { type: string }
        grams: { $ref: "#/components/schemas/Decimal" }
        is_default: { type: boolean }
        source: { $ref: "#/components/schemas/ServingSource" }
        sort_order: { type: integer, format: int32 }
    ServingCreate:
      type: object
      required: [label, grams]
      properties:
        label: { type: string }
        grams: { $ref: "#/components/schemas/Decimal" }
        is_default: { type: boolean, default: false }
        source:
          allOf:
            - $ref: "#/components/schemas/ServingSource"
          description: Defaults to `user` if omitted.
        sort_order: { type: integer, format: int32, default: 0 }
    ServingPatch:
      type: object
      description: |
        Partial update of a serving. Toggling `is_default` is done via the
        dedicated `POST /servings/{id}/default` endpoint, not this one.
      properties:
        label: { type: string }
        grams: { $ref: "#/components/schemas/Decimal" }
        sort_order: { type: integer, format: int32 }
    ServingPreview:
      type: object
      required: [id, label, grams]
      properties:
        id: { type: string, format: uuid }
        label: { type: string }
        grams: { $ref: "#/components/schemas/Decimal" }
    FoodDetail:
      type: object
      required:
        - id
        - source
        - name
        - categories_tags
        - nutrition
        - quality_score
        - servings
      properties:
        id: { type: string, format: uuid }
        source: { $ref: "#/components/schemas/FoodSource" }
        owner_user_id:
          type: [string, "null"]
          format: uuid
        barcode:
          type: [string, "null"]
        fdc_id:
          type: integer
          format: int64
          description: USDA FoodData Central id, present for `usda`-sourced foods only.
        name: { type: string }
        brands:
          type: [string, "null"]
        categories_tags:
          type: array
          items: { type: string }
        nutrition: { $ref: "#/components/schemas/NutritionPer100g" }
        nutriscore:
          oneOf:
            - $ref: "#/components/schemas/NutriscoreGrade"
            - type: "null"
        quality_score: { type: integer, format: int32 }
        extra_nutrients:
          description: Free-form JSON of additional nutrients, if any.
        servings:
          type: array
          items: { $ref: "#/components/schemas/Serving" }
    FoodCreate:
      type: object
      required: [name, nutrition]
      description: |
        Create a new user-custom food. The name `__quick_add__` is reserved
        for the internal quick-add sentinel and returns 400.
      properties:
        name: { type: string }
        brands: { type: string }
        barcode: { type: string }
        categories_tags:
          type: array
          items: { type: string }
          default: []
        nutrition: { $ref: "#/components/schemas/NutritionPer100g" }
        nutriscore_grade: { $ref: "#/components/schemas/NutriscoreGrade" }
...
    FoodPatch:
      type: object
      description: Any omitted field is left unchanged.
      properties:
        name: { type: string }
        brands: { type: string }
        barcode: { type: string }
        categories_tags:
          type: array
          items: { type: string }
        nutrition: { $ref: "#/components/schemas/NutritionPer100g" }
        nutriscore_grade: { $ref: "#/components/schemas/NutriscoreGrade" }
    FoodSearchHit:
      type: object
      required: [id, source, name]
      properties:
        id: { type: string, format: uuid }
        source: { $ref: "#/components/schemas/FoodSource" }
        name: { type: string }
        brand:
          type: [string, "null"]
        barcode:
          type: [string, "null"]
        default_serving:
          oneOf:
            - $ref: "#/components/schemas/ServingPreview"
            - type: "null"
        calories_per_serving:
          $ref: "#/components/schemas/NullableDecimal"
    Paginated:
      type: object
      required: [results, total, limit, offset]
      properties:
        results:
          type: array
          description: Replaced by the concrete schema in allOf compositions.
          items: {}
        total:
          type: integer
          format: int64
          description: Total matching rows across all pages.
        limit:
          type: integer
          format: int64
        offset:
          type: integer
          format: int64
    PaginatedFoodSearchHits:
      allOf:
        - $ref: "#/components/schemas/Paginated"
        - type: object
          properties:
            results:
              type: array
              items: { $ref: "#/components/schemas/FoodSearchHit" }
...